updated eslint rules
Turned off 'vue/require-explicit-emits' rule as this seems to be a Javascript rule that does not work well to Typescript. Other linter rules already flag undeclared emits (so this rule is redundant) and it does not understand the Typescript generic type arguement way of specifying emit type (so it is unhelpful).