Removed movable.color switching
I discovered that Chessgrounds only allows the color whose turn it is to move when `movable.color` is set to 'both', so theres actually no need to switch it back and forth between 'white' and 'black' manually every turn.

This kind of makes the `player-color` prop a bit pointless, as it turns out to just be a proxy for a config value. Instead of removing it I've kept it for backwards compatibility and simply merge its value into the config during `setConfig`.